# docs/swagger.yaml
basePath: /
definitions:
...
  request.RondaGroupCreateRequest:
    properties:
      name:
        type: string
    type: object
  request.RondaGroupMemberCreateRequest:
    properties:
      group_id:
        type: string
      resident_id:
        type: string
    type: object
  request.RondaGroupMemberUpdateRequset:
    properties:
      group_id:
        type: string
      id:
        type: string
      resident_id:
        type: string
    type: object
  request.RondaGroupUpdateRequset:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  utils.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      sort:
        type: string
      sort_by:
        type: string
      total_pages:
        type: integer
      total_rows:
        type: integer
    type: object
  utils.ResponseData:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/utils.Pagination'
      status:
        type: string
    type: object
host: localhost:5001
info:
  contact:
    email: sarikhin@yahoo.co.id
    name: API Support
  description: This is a API Server for SIRNAWA
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SIRNAWA API
  version: "1.0"
paths:
...
  /ronda-group:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Get List RondaGroup
      tags:
      - RondaGroup
    post:
      consumes:
      - application/json
      parameters:
      - description: Create RondaGroup
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RondaGroupCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Create RondaGroup
      tags:
      - RondaGroup
  /ronda-group-member:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Get List RondaGroupMember
      tags:
      - RondaGroupMember
    post:
      consumes:
      - application/json
      parameters:
      - description: Create RondaGroupMember
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RondaGroupMemberCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Create RondaGroupMember
      tags:
      - RondaGroupMember
  /ronda-group-member/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: RondaGroupMember id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Delete RondaGroupMember By ID
      tags:
      - RondaGroupMember
    get:
      consumes:
      - application/json
      parameters:
      - description: RondaGroupMember id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Find RondaGroupMember By ID
      tags:
      - RondaGroupMember
    put:
      consumes:
      - application/json
      parameters:
      - description: Update RondaGroupMember
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RondaGroupMemberUpdateRequset'
      - description: RondaGroupMember id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Update RondaGroupMember
      tags:
      - RondaGroupMember
  /ronda-group-member/paginated:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Order by
        in: query
        name: order_by
        type: string
      - description: Order
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Get Paginated RondaGroupMember
      tags:
      - RondaGroupMember
  /ronda-group/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: RondaGroup id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Delete RondaGroup By ID
      tags:
      - RondaGroup
    get:
      consumes:
      - application/json
      parameters:
      - description: RondaGroup id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Find RondaGroup By ID
      tags:
      - RondaGroup
    put:
      consumes:
      - application/json
      parameters:
      - description: Update RondaGroup
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RondaGroupUpdateRequset'
      - description: RondaGroup id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Update RondaGroup
      tags:
      - RondaGroup
  /ronda-group/paginated:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Order by
        in: query
        name: order_by
        type: string
      - description: Order
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Get Paginated RondaGroup
      tags:
      - RondaGroup
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"